modelando banco de dados e movendo estado de ligado ou desligado do bot indvidual da collection 'Chat' para 'Contact' adicionando campo dentro da collection Account.bot.activated para controlar o estado do bot entre ligado e desligado para todos os chats adicionando evento emit 'bot:get_info_bot' do websocket para buscar informacoes do bot adicionando evento emit 'bot:update_status_bot' do websocket para ligar e desligar o bot adicionando evento emit 'bot:update_prompt' do websocket para salvar o novo prompt da IA padronizando algumas respostas do back end (ou piorando ainda mais o padrao kkkkk vamos ver o que a pressa vai me permitir) refatorando codigo corrigindo bugs

+++ Socket/methods/on/spreadsheets/spreadsheet.js
@@ -9,19 +9,22 @@ import mongodb from "../../../../MongoDB/Mongodb.js";
  * @param {Object} callback FUNCAO DE RESPOSTA
 */
 export async function getSpreadsheets(socket, data, callback) {
-	const { idPhone } = socket.account;
+	const { idPhone  } = socket.account;
 
 	try {
 		const pages = await googleSheets.getPages(socket.account);
-		const account = await mongodb.Account.findOne({ idPhone: idPhone }).select("googleSheets.pages -_id");
+		const account = await mongodb.Account.findOne({ idPhone: idPhone }).select("googleSheets -_id");
 
 setTimeout(() => {
-		callback(pages.map((page) => {
-			return ({
-				page: page,
-				active: account.googleSheets.pages.includes(page)
-			});
-		}));
+		callback({
+			url: account.googleSheets.spreadsheet,
+			pages: pages.map((page) => {
+				return ({
+					page: page,
+					active: account.googleSheets.pages.includes(page)
+				});
+			})
+		});
 }, 1000);
 	} catch (error) {
 		await mongodb.saveError(idPhone, `Error no metodo "getSpreadsheets": ${error}`);
@@ -41,10 +44,11 @@ export async function updateUsedSpreadsheets(socket, data, callback) {
 
 	try {
 		if ((add && remove) || (!add && !remove)) return (callback({ error: "Deve haver uma ação de adicionar ou remover planilha" }));
+		if ((add && typeof add !== "string") || (remove && typeof remove !== "string")) return (callback({ error: 'O campo "add" ou "remove" deve ser do tipo string' }));
 		if (add) await mongodb.addSpreadsheets(idPhone, add);
 		if (remove) await mongodb.removeSpreadsheets(idPhone, remove);
 setTimeout(() => {
-		callback(200);
+		callback((add) ? { add: add } : { remove: remove });
 }, 1000);
 	} catch (error) {
 		await mongodb.saveError(idPhone, `Error no metodo "updateUsedSpreadsheets": ${error}`);
